requirements.txt
amqp==5.0.2
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1
django-allauth==0.44.0
django-cors-headers==3.6.0
django-crontab==0.7.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
oauthlib==3.1.0
prompt-toolkit==3.0.10
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
